+++ changelog.md
@@ -1,5 +1,6 @@
 ﻿## 5.0.2
 - Debug: remove console.log for polls monitoring ([#202](https://github.com/Eccleria/ewibot/pull/202))
+- Feat: refacto on activities ([#204](https://github.com/Eccleria/ewibot/pull/204))
 - Fix: typo on anonymous check during poll refresh ([#199](https://github.com/Eccleria/ewibot/pull/199))
 
 ## 5.0.1